Moved installer handler unit and functional tests into own classes.
