+++ .vortex/installer/tests/Functional/Handlers/AiCodeInstructionsInstallTest.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DrevOps\VortexInstaller\Tests\Functional\Handlers;
+
+use DrevOps\VortexInstaller\Prompts\Handlers\AiCodeInstructions;
+use DrevOps\VortexInstaller\Prompts\PromptManager;
+use DrevOps\VortexInstaller\Utils\Env;
+use PHPUnit\Framework\Attributes\CoversClass;
+
+#[CoversClass(AiCodeInstructions::class)]
+class AiCodeInstructionsInstallTest extends AbstractInstallTestCase {
+
+  public static function dataProviderInstall(): array {
+    return [
+      'ai instructions, claude' => [
+        static::cw(fn() => Env::put(PromptManager::makeEnvName(AiCodeInstructions::id()), AiCodeInstructions::CLAUDE)),
+      ],
+
+      'ai instructions, none' => [
+        static::cw(fn() => Env::put(PromptManager::makeEnvName(AiCodeInstructions::id()), AiCodeInstructions::NONE)),
+      ],
+    ];
+  }
+
+}
