+++ .vortex/installer/tests/Unit/Handlers/DatabaseDownloadSourcePromptManagerTest.php
@@ -1,0 +1,42 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DrevOps\VortexInstaller\Tests\Unit\Handlers;
+
+use DrevOps\VortexInstaller\Prompts\Handlers\DatabaseDownloadSource;
+use PHPUnit\Framework\Attributes\CoversClass;
+use Laravel\Prompts\Key;
+
+#[CoversClass(DatabaseDownloadSource::class)]
+class DatabaseDownloadSourcePromptManagerTest extends AbstractPromptManagerTestCase {
+
+  public static function dataProviderRunPrompts(): array {
+    $expected_defaults = static::getExpectedDefaults();
+
+    return [
+      'database download source - prompt' => [
+        [DatabaseDownloadSource::id() => Key::ENTER],
+        [DatabaseDownloadSource::id() => DatabaseDownloadSource::URL] + $expected_defaults,
+      ],
+
+      'database download source - discovery' => [
+        [],
+        [DatabaseDownloadSource::id() => DatabaseDownloadSource::FTP] + $expected_defaults,
+        function (AbstractPromptManagerTestCase $test): void {
+          $test->stubDotenvValue('VORTEX_DB_DOWNLOAD_SOURCE', DatabaseDownloadSource::FTP);
+        },
+      ],
+
+      'database download source - discovery - invalid' => [
+        [],
+        $expected_defaults,
+        function (AbstractPromptManagerTestCase $test): void {
+          $test->stubDotenvValue('VORTEX_DB_DOWNLOAD_SOURCE', 'invalid_source');
+        },
+      ],
+
+    ];
+  }
+
+}
